Do not add the key fields if the cache plugin is present already
