+++ libraries/apollo-compiler/src/test/kotlin/com/apollographql/apollo/compiler/keyfields/KeyFieldsTest.kt
@@ -2,50 +2,16 @@ package com.apollographql.apollo.compiler.keyfields
 
 import com.apollographql.apollo.ast.GQLDocument
 import com.apollographql.apollo.ast.GQLExecutableDefinition
-import com.apollographql.apollo.ast.GQLFragmentDefinition
-import com.apollographql.apollo.ast.GQLOperationDefinition
 import com.apollographql.apollo.ast.parseAsGQLDocument
 import com.apollographql.apollo.ast.toGQLDocument
 import com.apollographql.apollo.ast.validateAsSchemaAndAddApolloDefinition
-import com.apollographql.apollo.compiler.internal.addRequiredFields
-import com.apollographql.apollo.compiler.internal.checkKeyFields
 import okio.Path.Companion.toPath
 import kotlin.test.Test
 import kotlin.test.assertContains
 import kotlin.test.assertEquals
 import kotlin.test.assertTrue
-import kotlin.test.fail
 
 class KeyFieldsTest {
-  @Test
-  fun testAddRequiredFields() {
-    val schema = "src/test/kotlin/com/apollographql/apollo/compiler/keyfields/schema.graphqls"
-        .toPath()
-        .toGQLDocument()
-        .validateAsSchemaAndAddApolloDefinition()
-        .getOrThrow()
-
-    val definitions = "src/test/kotlin/com/apollographql/apollo/compiler/keyfields/operations.graphql".toPath()
-        .toGQLDocument()
-        .definitions
-
-    val fragments = definitions.filterIsInstance<GQLFragmentDefinition>().associateBy { it.name }
-
-    val operation = definitions
-        .filterIsInstance<GQLOperationDefinition>()
-        .first()
-
-    try {
-      checkKeyFields(operation, schema, emptyMap())
-      fail("an exception was expected")
-    } catch (e: Exception) {
-      assertTrue(e.message?.contains("are not queried") == true)
-    }
-
-    val operationWithKeyFields = addRequiredFields(operation, "ifFragments", schema, fragments)
-    checkKeyFields(operationWithKeyFields, schema, emptyMap())
-  }
-
   @Test
   fun testExtendInterfaceTypePolicyDirective() {
     val schema = "src/test/kotlin/com/apollographql/apollo/compiler/keyfields/extendsSchema.graphqls"
