+++ tests/compiler-plugins/capitalize-enum-values/src/main/kotlin/hooks/TestPlugin.kt
@@ -1,3 +1,5 @@
+@file:Suppress("DEPRECATION")
+
 package hooks
 
 import com.apollographql.apollo.compiler.ApolloCompilerPlugin
