+++ tests/compiler-plugins/prefix-names/src/main/kotlin/hooks/TestPlugin.kt
@@ -1,8 +1,9 @@
+@file:Suppress("DEPRECATION")
+
 package hooks
 
 import com.apollographql.apollo.compiler.ApolloCompilerPlugin
 import com.apollographql.apollo.compiler.ApolloCompilerPluginEnvironment
-import com.apollographql.apollo.compiler.ApolloCompilerPluginLogger
 import com.apollographql.apollo.compiler.ApolloCompilerPluginProvider
 import com.apollographql.apollo.compiler.ApolloCompilerRegistry
 import com.apollographql.apollo.compiler.CodegenSchema
