+++ tests/compiler-plugins/schema-codegen/src/main/kotlin/schema/TestPlugin.kt
@@ -1,22 +1,21 @@
+@file:Suppress("DEPRECATION")
+
 package schema
 
 import com.apollographql.apollo.ast.ForeignSchema
 import com.apollographql.apollo.ast.GQLIntValue
 import com.apollographql.apollo.ast.GQLTypeDefinition
-import com.apollographql.apollo.ast.Schema
 import com.apollographql.apollo.ast.parseAsGQLDocument
 import com.apollographql.apollo.compiler.ApolloCompilerPlugin
 import com.apollographql.apollo.compiler.ApolloCompilerPluginEnvironment
 import com.apollographql.apollo.compiler.ApolloCompilerPluginLogger
 import com.apollographql.apollo.compiler.ApolloCompilerPluginProvider
 import com.apollographql.apollo.compiler.ApolloCompilerRegistry
-import com.apollographql.apollo.compiler.SchemaListener
 import com.squareup.kotlinpoet.ClassName
 import com.squareup.kotlinpoet.FileSpec
 import com.squareup.kotlinpoet.PropertySpec
 import okio.buffer
 import okio.source
-import java.io.File
 
 class TestPlugin(
     logger: ApolloCompilerPluginLogger,
